cambios de dto para mantener logica de la db por ensima del mapeo de codigo directo de entty , se sanea basatnte cosas puede que aparezcan malfuncionamientos pero es necesario

+++ GotsThorlabs/Interfaces/ITourCrudService.cs
@@ -6,9 +6,9 @@ namespace GotsThorlabs.Interfaces
     public interface ITourCrudService
     {
         Task<IEnumerable<TourEntity>> GetAllAsync(CancellationToken ct);
-        Task<TourEntity?> GetByIdAsync(int id, CancellationToken ct);
+        Task<TourEntity?> GetByIdAsync(long id, CancellationToken ct);
         Task<TourEntity> CreateAsync(TourDTO dto, CancellationToken ct);
-        Task UpdateAsync(int id, TourDTO dto, CancellationToken ct);
-        Task DeleteAsync(int id, CancellationToken ct);
+        Task UpdateAsync(long id, TourDTO dto, CancellationToken ct);
+        Task DeleteAsync(long id, CancellationToken ct);
     }
 }
